mailbox: remove the header from struct mailbox_msg
The mailbox_msg is restricted to just two elements, a size field and
a data pointer containing the data.

NOTE: The size field is retained for now in the mailbox_msg structure
for using the kfifos. The deciding factor would be the mechanism used
for buffering. The current kfifo usage implies that the users need not
maintain a reference to a queued buffer (obviously, we are expecting
the transport packets to be small here). If the size field is removed,
then the users do have to maintain the buffers, and the common code
can just maintain the pointers.

BIG NOTE: Will impact any out-of-tree DBX500 code, due to the structure
change.

Signed-off-by: Suman Anna <s-anna@ti.com>