mailbox: descope the exported functions for interrupts & context
The four exported API that were currently dealing with enabling/disabling
interrupts and saving/restoring contexts have been moved into an OMAP-
specific header file, and removed from the public exported API.

These are needed to be exported for the state machine in TI DSP/Bridge
driver. The save_context and restore_context can be eliminated with
usage similar to remoteproc driver and runtime_pm support, but the
other two may need to be exposed to keep the TI DSP/Bridge driver
functional.

Signed-off-by: Suman Anna <s-anna@ti.com>