TEMP: mailbox/omap: add support for AM335x wkupm3 mbox
The WakeM3 can only receive the interrupt in AM335x, but cannot
read the mailbox message. The MPU needs to read back the message.
The current solution writes the message and waits for 100us before
reading back the message, since we do not want to expose a new
OMAP specific API for dealing with this.

Signed-off-by: Suman Anna <s-anna@ti.com>