TEMP: mailbox/omap: fix driver for wkup_m3 suspend
This is a temporary staging patch for achieving the
AM335 suspend functionality. The changes rely on
using the same Tx and Rx ids for the wkup_m3 mailbox
and adjusting the driver logic to use the MPU user
id for dealing with Tx-ready interrupts, and the
WkupM3 user id for dealing with Rx interrupts.

The changes themselves are hacky and the code needs
to be fixed up to use separate link ops for AM335
Wkup_M3 for better streamlined code.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>