mailbox/omap: add code to support the wkupm3 operations
The WkupM3 mailbox used for triggering PM operations such as suspend
and resume on AM33x/AM43x is special in that the M3 processor cannot
access the mailbox registers. However, an interrupt is needed to be
sent to request the M3 to perform a desired PM operation. This patch
adds the support for this special mailbox through separate ops for
this mailbox.

NOTE: The current patch assigns the new controller ops for all links
      on AM33x/AM43x that uses a WkupM3 mailbox with logic adjusted
      for non-WkupM3 mailboxes. This may be a bit inelegant, but a
      newer version of mailbox framework is expected to have ops per
      link, which should improve the code a bit.

Signed-off-by: Suman Anna <s-anna@ti.com>