mailbox/omap: add code to support the wkupm3 operations
The WkupM3 mailbox used for triggering PM operations such as suspend
and resume on AM33x/AM43x is special in that the M3 processor cannot
access the mailbox registers. However, an interrupt is needed to be
sent to request the M3 to perform a desired PM operation. This patch
adds the support for this special mailbox through separate ops for
this mailbox.

NOTE: This patch is done prior to the OMAP adaptation to the mailbox
      framework, and can be easily adapted. Remove this comment if
      posting upstream.

Signed-off-by: Suman Anna <s-anna@ti.com>