TST: arm64: dts: ti: k3-am654-base-board: Add mailbox loopback devices for testing
Add a loopback sub-mailbox device within Mailbox clusters 1 and 2 on the
AM65x SoCs for testing. The loopback device uses the same mailbox fifo
for sending and receiving messages on the MPU itself, facilitating a
unit-test. The last FIFO within the IP is chosen for the loopback to
not conflict with the actual FIFOs at the beginning that will be used
to communicate with the remote processors.

Signed-off-by: Suman Anna <s-anna@ti.com>