TST: ARM: dts: OMAP2+: Add mailbox loopback devices for testing
Add loopback mailbox devices on all devices newer than OMAP4
(OMAP4, OMAP5, AM33xx, AM43xx and DRA7) for testing. The loopback
devices uses a same mailbox fifo for sending and receiving messages
on the MPU itself, facilitating a unit-test. The MPU utilizes both
the Tx and Rx interrupts as part of the message communications.

The following are the mailbox fifos used on different devices:
	OMAP4/OMAP5: 4
	AM33x/AM43x: 5
	DRA7: 0 on mailboxes #4 and #5,
              6 on remaining mailboxes

Signed-off-by: Suman Anna <s-anna@ti.com>